# internal/api/openapi/swagger.yaml
basePath: /api
definitions:
  api.APIError:
    properties:
      code:
        type: integer
      details:
        additionalProperties:
          type: string
        type: object
      error:
        type: string
    type: object
...
  api.AthleteNote:
    properties:
      athlete_id:
        type: integer
      author:
        type: string
      author_id:
        type: integer
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      is_private:
        type: boolean
      pinned:
        type: boolean
      updated_at:
        type: string
    type: object
  api.AthleteNoteRequest:
    properties:
      content:
        type: string
      is_private:
        type: boolean
      pinned:
        type: boolean
    type: object
...
  api.ExerciseHistoryDay:
    properties:
      sets:
        items:
          $ref: '#/definitions/api.ExerciseHistoryEntry'
        type: array
      workout_date:
        type: string
      workout_id:
        type: integer
    type: object
  api.ExerciseHistoryEntry:
    properties:
      notes:
        type: string
      reps:
        type: integer
      rpe:
        type: number
      set_number:
        type: integer
      weight:
        type: number
      workout_date:
        type: string
      workout_id:
        type: integer
    type: object
  api.ExerciseHistoryPage:
    properties:
      days:
        items:
          $ref: '#/definitions/api.ExerciseHistoryDay'
        type: array
      has_more:
        type: boolean
    type: object
...
  api.GoalRequest:
    properties:
      goal:
        type: string
    type: object
...
  api.PasskeyLabelRequest:
    properties:
      label:
        example: My YubiKey
        type: string
    type: object
  api.PasskeyResponse:
    properties:
      backup_state:
        type: boolean
      created_at:
        type: string
      id:
        type: integer
      label:
        type: string
      last_used_at:
        type: string
      use_count:
        type: integer
    type: object
...
  api.StatusResponse:
    properties:
      status:
        example: ok
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    REST API for the RepLog workout tracking application.

    Authentication uses session cookies issued by `POST /api/login`
    (or the WebAuthn / token-login endpoints). Cookies are
    `HttpOnly`, `SameSite=Lax`, and (in production) `Secure`.
    The browser sends them automatically on subsequent requests.
  title: RepLog API
  version: 1.0.0
paths:
...
  /athletes/{id}/exercises/{exerciseID}/history:
    get:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Exercise ID
        in: path
        name: exerciseID
        required: true
        type: integer
      - description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ExerciseHistoryPage'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Exercise history for athlete
      tags:
      - Exercises
...
  /athletes/{id}/goal:
    put:
      consumes:
      - application/json
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Goal
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.GoalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Update athlete goal
      tags:
      - Athletes
...
  /athletes/{id}/notes:
    post:
      consumes:
      - application/json
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Note
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AthleteNoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.AthleteNote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Create athlete journal note
      tags:
      - Athletes
  /athletes/{id}/notes/{noteID}:
    delete:
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Note ID
        in: path
        name: noteID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Delete athlete note
      tags:
      - Athletes
    put:
      consumes:
      - application/json
      parameters:
      - description: Athlete ID
        in: path
        name: id
        required: true
        type: integer
      - description: Note ID
        in: path
        name: noteID
        required: true
        type: integer
      - description: Note
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.AthleteNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.AthleteNote'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.APIError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Update athlete note
      tags:
      - Athletes
...
  /passkeys:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.PasskeyResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIError'
      summary: List user's passkeys
      tags:
      - Passkeys
  /passkeys/{id}:
    delete:
      description: Idempotent — returns 200 even if the credential is already gone
        or belongs to another user (does NOT leak existence).
      parameters:
      - description: Credential ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Delete passkey
      tags:
      - Passkeys
  /passkeys/label:
    post:
      consumes:
      - application/json
      description: Called by the SPA right before BeginRegistration so the label is
        paired to the credential being created.
      parameters:
      - description: Label
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.PasskeyLabelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.StatusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.APIError'
      summary: Stash passkey label for the in-flight ceremony
      tags:
      - Passkeys
...
swagger: "2.0"
...